// front/src/services/generated/models/ProductionPreset.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * 可执行的固定工作流预设；不接受任意步骤图。
 */
export type ProductionPreset = 'script_assist' | 'prepare_shots' | 'prepare_frames' | 'controlled_video';